LayoutTests/platform/chromium-linux/plugins/mouse-events-expected.txt

CONSOLE MESSAGE: line 16: PLUGIN: adjustCursorEvent
CONSOLE MESSAGE: line 16: PLUGIN: adjustCursorEvent
CONSOLE MESSAGE: line 17: PLUGIN: getFocusEvent
CONSOLE MESSAGE: line 17: PLUGIN: mouseDown at (12, 12)
CONSOLE MESSAGE: line 18: PLUGIN: mouseUp at (12, 12)
CONSOLE MESSAGE: line 19: PLUGIN: adjustCursorEvent
CONSOLE MESSAGE: line 20: PLUGIN: mouseDown at (22, 22)
CONSOLE MESSAGE: line 22: PLUGIN: adjustCursorEvent
CONSOLE MESSAGE: line 22: PLUGIN: mouseUp at (32, 22)
CONSOLE MESSAGE: line 23: PLUGIN: adjustCursorEvent

Test for bug 11517: Flash clicks/interactivity not working properly.
